requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
aiofiles==23.2.1
python-multipart==0.0.6
supabase==1.2.0
requests==2.31.0
openai==1.3.5
python-dotenv==1.0.0
python-dateutil==2.8.2